requirements.txt
<<<<<<< HEAD
click==7.1.1
Flask==1.1.1
=======
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
gunicorn==20.0.4
idna==2.9
>>>>>>> 2627d30b5fcd194d2fbabe699d0bb772ab629086
itsdangerous==1.1.0
Jinja2==2.11.1
jsonify==0.5
MarkupSafe==1.1.1
<<<<<<< HEAD
<<<<<<< HEAD
=======
=======
>>>>>>> 2627d30b5fcd194d2fbabe699d0bb772ab629086
pytube3==9.6.4
requests==2.23.0
typing-extensions==3.7.4.1
urllib3==1.25.8
<<<<<<< HEAD
>>>>>>> d5879b3... New Commit part 1
=======
>>>>>>> 2627d30b5fcd194d2fbabe699d0bb772ab629086
Werkzeug==1.0.0
